requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pillow
google-genai
